fix: Add Sync and Send bounds to Errors
There errors do not need to be so restrictive, by restricting them, we
create a lot more freedom in the actual logic we are writing.

This might bite us later, but as far as I can see we Errors can be
completely owned values, and borrowing from the application, puts
constraints on the application where that is not necessary.